+++ src/app/core/http/offer/offer.service.ts
@@ -1,14 +1,14 @@
 import { Injectable } from '@angular/core';
-import {
-  BehaviorSubject,
-  Observable
-} from 'rxjs';
+import { BehaviorSubject, forkJoin, Observable } from 'rxjs';
 import { tap } from 'rxjs/operators';
 
 import { ApiService } from 'src/app/core/http/api/api.service';
+import { OfferPropertyCache } from 'src/app/core/http/api/offer-property-cache.service';
+
 import { OfferToAPI } from 'src/app/core/http/api/api.interfaces';
 import { Offer, PartialOffer } from 'src/app/core/models/offer';
-import { LOAD, ADD, EDIT, REMOVE, OfferStore } from './offer.store';
+import { LOAD, ADD, EDIT, REMOVE, OfferStore } from 'src/app/core/http/store/offer.store';
+import { OfferPropertyList } from 'src/app/core/models/offer-properties';
 
 /**
  * offer.service.ts
@@ -24,42 +24,71 @@ export class OfferService {
   offers$: BehaviorSubject<Offer[]>;
   //offerChanged$ = new BehaviorSubject({});
 
-  constructor(private apiService: ApiService, private offerStore: OfferStore) {
+  constructor(
+    private apiService: ApiService,
+    private offerStore: OfferStore,
+    private offerPropertyCache: OfferPropertyCache
+  ) {
     this.offers$ = offerStore.items$;
   }
 
-  getAllOffers(): Observable<Offer[]> {
-    this.apiService
-      .getAllOffers()
+  /**
+   * preloads OfferShortList (OfferTiles)
+   * only call once in data-handler
+   * @returns
+   */
+  preloadAllOfferShortList(): any {
+    const property$ = this.offerPropertyCache.loadOfferProperties();
+    const shortOffer$ = this.offerPropertyCache.loadShortOfferList();
+
+    // Parallel laden, aber erst auswerten wenn beide completed sind
+    forkJoin([shortOffer$, property$])
       .pipe(
-        tap((offers) => {
+        tap((results) => {
+          let offers = this.mapDataInOfferStructure(results[0]);
           this.offerStore.dispatch({ type: LOAD, data: offers });
         })
       )
       .subscribe(
         (offers) => {
-          //console.log('Offers geladen', offers);
+          console.log('Offers(short) geladen', offers);
         },
         (error) => {
-          console.log('getAllOffer_Error:', error);
+          console.log('getAllOfferShortList_error:', error);
           let message = error; // Anpassen wenn nötig
           this.offers$.error(message);
+        },
+        () => {
+          console.log('getAllOfferShortList completed');
         }
       );
+
     return this.offers$;
   }
 
-  getAllOfferShortList(): Observable<PartialOffer[]> {
-    this.apiService
-      .getAllOffers()
+  /**
+   * Gets Offer Structure with all Data
+   * only call once in data-handler
+   * @deprecated Use getAllOfferShortList
+   * @returns
+   */
+   preloadAllOffers(): Observable<Offer[]> {
+    const property$ = this.offerPropertyCache.loadOfferProperties();
+    const longOffer$ = this.offerPropertyCache.loadLongOfferList();
+
+    forkJoin([longOffer$, property$])
       .pipe(
-        tap((offers) => {
+        tap((results) => {
+          let offers = results[0];
+          for (const offerItem of offers) {
+            offerItem.competence_text = this.createCompetenceString(offerItem.competences);
+          }
           this.offerStore.dispatch({ type: LOAD, data: offers });
         })
       )
       .subscribe(
         (offers) => {
-          //console.log('Offers geladen', offers);
+          console.log('Offers(long) geladen', offers);
         },
         (error) => {
           console.log('getAllOffer_Error:', error);
@@ -79,6 +108,7 @@ export class OfferService {
       tap((savedOffer) => {
         console.log('New Offer:', savedOffer);
         // this.offerChanged$.next(savedOffer);
+        savedOffer.competence_text = this.createCompetenceString(savedOffer.competences);
         const action = { type: ADD, data: savedOffer };
         this.offerStore.dispatch(action);
       })
@@ -86,13 +116,13 @@ export class OfferService {
   }
 
   // Angedacht für die Sortierungsliste.
-  storePartialOfferList(offerList: PartialOffer[]){
+  storePartialOfferList(offerList: PartialOffer[]) {
     const sentList = offerList.filter((offer) => offer.id !== null);
     return this.apiService.updatePartialOfferList(sentList).pipe(
-      tap(savedOffers => {
+      tap((savedOffers) => {
         // TODO: Einzelne Datensätze updaten oder besser die Datensätze neu laden
       })
-    )
+    );
   }
 
   updateOffer(id: number, data: OfferToAPI) {
@@ -101,6 +131,7 @@ export class OfferService {
       tap((savedOffer) => {
         // this.offerChanged$.next(savedOffer);
         console.log('Updated Offer:', savedOffer);
+        savedOffer.competence_text = this.createCompetenceString(savedOffer.competences);
         const action = { type: EDIT, data: savedOffer };
         this.offerStore.dispatch(action);
       })
@@ -120,4 +151,58 @@ export class OfferService {
   clearStoreData() {
     this.offerStore.dispatch({ type: LOAD, data: null });
   }
+
+
+  public getOfferProperties(): Observable<Map<string, OfferPropertyList>>{
+    return this.offerPropertyCache.loadOfferProperties();
+  }
+
+  public createCompetenceString(competences: number[]) : string {
+    let competenceArr = [];
+    let competenceStr = '';
+
+    if (competences.length == 0) {
+      competenceStr = 'keine Angabe';
+    } else {
+
+      for (const competence of competences) {
+        const competenceText = this.offerPropertyCache.competencesMap.get(competence);
+        competenceArr.push(this.compworkaroundForText(competenceText));
+      }
+      competenceStr = competenceArr.join(', ');
+    }
+    return competenceStr;
+  }
+
+  /**
+   * Die Funktion sorgt dafür dass die Kompetenztexte so angezeigt werden wie vorher
+   * Todo: Besser die einheitliche Description benutzen (siehe offer-property-cache.service)
+   * Die ist im Moment recht lang (siehe Texte in den Filtern)
+   * @param text
+   * @returns
+   */
+  private compworkaroundForText(text: string): string {
+    switch(text){
+      case 'tech': return 'Tech';
+      case 'digital': return 'Digital Basic';
+      case 'classic': return 'Classic';
+      default: return 'NA';
+    }
+  }
+
+  mapDataInOfferStructure(offers: Partial<Offer>[]): Partial<Offer>[] {
+    for (const offerItem of offers) {
+      offerItem.institution = {
+        id: offerItem.institution_id,
+        title: this.offerPropertyCache.institutionMap.get(offerItem.institution_id),
+        url: undefined,
+      };
+      offerItem.language = this.offerPropertyCache.languageMap.get(offerItem.language_id);
+      offerItem.type = this.offerPropertyCache.formatMap.get(offerItem.offertype_id);
+      offerItem.competence_text = this.createCompetenceString(offerItem.competences);
+    }
+    return offers;
+  }
+
+
 }