+++ Changelog.md
@@ -2,6 +2,9 @@
 
 ## [Unreleased]
 
+## 2021-06-02/ml
+- Offer-Object (short) is created out of Offer-Short + Properties
+- generated competence-text
 
 ## 2021-05-27/ml
 # Added